test: consolidate E2E test suite and address review feedback
- Consolidate all E2E integration tests into McpToolboxClientE2ETest
- Add text() convenience method to ToolResult
- Add tests for null parameters, type mismatches, and invalid map types
- Use Jackson structural JSON node parsing in process-data assertions
- Use destroyForcibly() for server process teardown
- Parameterize protocol version test across supported versions

TAG=agy
CONV=f9ec8a85-66e3-479e-a944-3b621f8bcd18
